User CRUD and authenticator separated
