+++ www/Controller/User.class.php
@@ -2,70 +2,11 @@
 
 namespace App\Controller;
 
-use App\Core\Authenticator;
-use App\Core\Verificator;
 use App\Core\View;
 use App\Model\User as UserModel;
 
 class User
 {
-    public function login()
-    {
-        $user = new UserModel();
-        $view = new View("login");
-        $view->assign("user", $user);
-
-        if (!empty($_POST)) {
-            // TO DO create a find
-            extract($_POST);
-            $errors = Verificator::checkForm($user->getLoginForm(), $_POST);
-            if (count($errors) == 0) {
-                $auth = new Authenticator();
-                $auth->login($_POST);
-            }
-        }
-    }
-
-
-    public function register()
-    {
-
-        $user = new UserModel();
-        if (!empty($_POST)) {
-
-            //Je vérifie qu'il les entrées soient corrects
-            $errors = Verificator::checkForm($user->getRegisterForm(), $_POST);
-            if (count($errors) === 0) {
-
-                $user->setFirstname($_POST['firstname']);
-                $user->setLastname($_POST['lastname']);
-                $user->setEmail($_POST['email']);
-                $user->setPassword($_POST['password']);
-                $user->save();
-            } else {
-                print_r($errors);
-            }
-        }
-
-        $view = new View("register");
-        $view->assign("user", $user);
-    }
-
-
-    public function logout()
-    {
-        $auth = new Authenticator();
-
-        $view = new View("logout");
-        $view->assign("auth", $auth);
-    }
-
-
-    public function pwdforget()
-    {
-        echo "Mot de passe oublié";
-    }
-    
     /** function index pour avoir la liste de tous les users présents en bdd */
     public function index()
     {
@@ -73,7 +14,7 @@ class User
         $userTable = "esgi_user";
         $users = $user->findAll($userTable);
 
-        $view = new View("viewUser");
+        $view = new View("viewUsers");
         $view->assign("users", $users);
     }
 
@@ -86,7 +27,6 @@ class User
 
         $view = new View("showUser");
         $view->assign("user", $users);
-
     }
 
     /** function update pour modifier les informations d'un user spécifique */
@@ -98,7 +38,6 @@ class User
 
         $view = new View("editUser");
         $view->assign("user", $user);
-
     }
 
     public function update()
@@ -112,16 +51,15 @@ class User
 
         $user->save();
 
-        header("Location: /user/index");
+        header("Location: /user/show?id=$userId");
     }
 
     /** function delete pour supprimer un user spécifique */
     public function delete()
     {
         $user = new UserModel();
-        $userId = $_GET['id'];
         $userTable = "esgi_user";
-        $users = $user->delete($userTable);
+        $user->delete($userTable);
 
         header("Location: /user/index");
     }
